requirements.txt
Django
mysql
django-import-export
django-font-awesome
django-bootstrap
django-bootstrap4
Pillow
xhtml2pdf
asgiref

django-crum
django-widget-tweaks

freeze

html5lib

mysqlclient
PyPDF2

pytz

reportlab

six
sqlparse

webencodings

xhtml2pdf